requirements.txt
amqp==2.4.2
billiard==3.6.0.0
blessings==1.7
bpython==0.18
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
curtsies==0.3.0
dj-database-url==0.5.0
Django>=2.1.7
django-extensions==2.1.7
greenlet==0.4.15
idna==2.8
kombu==4.5.0
psycopg2==2.8.2
Pygments==2.4.1
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.9
requests==2.21.0
six==1.12.0
typing==3.6.6
urllib3==1.24.3
vine==1.3.0
wcwidth==0.1.7
django-summernote
django-filebrowser==3.11.1
Pillow